+++ Engine/src/main/resources/templates/PhpLumen/Controller.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use DB;
 use Gate;
 use App\Utils\Arrays;
+use App\Utils\HttpStatus;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 
@@ -17,14 +18,19 @@ class {{resourceClass}}Controller extends RestController {
     // ----------------------------
 
 {{> Controller-resource}}{{! Indentation already in partial file }}
-
+{{> Controller-files}}
 
     // Relationship end-points callbacks
     // ---------------------------------
     {{#each relations}}
-    public static function get{{subresourceClassPlural}}(Request $req, $entityId) {
-        $entity = {{resourceClass}}::findOrFail($entityId);
+    {{#unless isBelongTo}}{{#unless isOneToOne}}public function get{{subresourceClassPlural}}(Request $req, $entityId) {
+        if (Gate::denies('read', new {{subresourceClass}}())) {
+            return self::abort(HttpStatus::Forbidden, [
+                'title' => 'Not allowed to read the resource'
+            ]);
+        }
 
+        $entity = {{resourceClass}}::findOrFail($entityId);
         $query = $entity->{{subFn}}();
 
         $query = self::addFieldsToQuery($query, $req);
@@ -41,31 +47,66 @@ class {{resourceClass}}Controller extends RestController {
             return {{subresourceClass}}::isFilterable($prop);
         });
 
-        $entities = $query->get();
+        $related = $query->get();
 
-        return response()->json($entities);
-    }
+        foreach ($related as &$ent) {
+            self::addHalLink($ent, 'self', $req->path() . '/' . $ent->id);
+        }
+        {{#if isManyToMany}}
+        foreach ($related as &$ent) {
+            $ent->_embedded = (object) [
+                'pivot' => $ent->pivot
+            ];
+            unset($ent->pivot);
+        }{{/if}}
 
-    public static function create{{subresourceClassPlural}}(Request $req, $entityId) {
+        $result = new \stdClass;
+        $result->count = count($related);
+        $result->total = $entity->{{subFn}}()->count();
+
+        self::addQueryMade($result, $req);
+        self::addHalPageLinks($result, $req, $result->total);
+        self::addHalEmbedded($result, '{{subresourceNamePlural}}', $related);
+
+        return response()->json($result);
+    }{{/unless}}{{/unless}}
+    {{#if isBelongTo}}
+{{> Controller-getSingleResourceFunction}}{{/if}}
+    {{#if isOneToOne}}
+{{> Controller-getSingleResourceFunction}}{{/if}}
+
+    public function create{{subresourceClassPlural}}(Request $req, $entityId) {
+        if (Gate::denies('create', new {{subresourceClass}}())) {
+            return self::abort(HttpStatus::Forbidden, [
+                'title' => 'Not allowed to create the resource'
+            ]);
+        }
+
         $input = $req->all();
 
         if (Arrays::isAssoc($input)) {
             $this->validate($req, {{subresourceClass}}::getValidationRules());
         } else if (is_array($input)) {
-            {{#if isOneToX}}if (count($input) > 1) {
-                // ...
+            {{#if isOneToOne}}if (count($input) > 1) {
+                return self::abort(HttpStatus::BadRequest, [
+                    'title' => 'Too many resources provided',
+                    'detail' => 'Only one resource representation is accepted. ' . count($input) . ' provided'
+                ]);
             }
             $validator = Validator::make($input[0], {{subresourceClass}}::getValidationRules());
             if ($validator->fails()) {
-                return self::error($validator->messages());
+                return self::abort(HttpStatus::UnprocessableEntity, ['invalid-fields' => $validator->messages()]);
             }{{else}}foreach ($input as $relatedEntity) {
                 $validator = Validator::make($relatedEntity, {{subresourceClass}}::getValidationRules());
                 if ($validator->fails()) {
-                    return self::error($validator->messages());
+                    return self::abort(HttpStatus::UnprocessableEntity, ['invalid-fields' => $validator->messages()]);
                 }
             }{{/if}}
         } else {
-            // ...
+            return self::abort(HttpStatus::BadRequest, [
+                'title' => 'Wrong data format',
+                'detail' => 'Only JSON objects and JSON arrays are valid'
+            ]);
         }
 
         $entity = {{resourceClass}}::findOrFail($entityId);
@@ -74,11 +115,36 @@ class {{resourceClass}}Controller extends RestController {
 
         return response()->json($relatedEntities);
     }
+    {{#if isManyToMany}}
+    public function getPivotWith{{subresourceClass}}($entityId, $relatedEntityId) {
+        if (Gate::denies('read', new {{resourceClass}}()) || Gate::denies('read', new {{subresourceClass}}())) {
+            return self::abort(HttpStatus::Forbidden, [
+                'title' => 'Not allowed to get the resource'
+            ]);
+        }
 
-    {{#if isManyToMany}}public static function unlink{{subresourceClass}}($entityId, $relatedEntityId) {
+        $pivot = (object) null;
         $entity = {{resourceClass}}::findOrFail($entityId);
         foreach ($entity->{{subFn}} as $relatedEntity) {
-            if ($relatedEntity->{{subresourceId}} === $relatedEntityId) {
+            if ($relatedEntity->id === (int) $relatedEntityId) {
+                $pivot = $relatedEntity->pivot;
+                break;
+            }
+        }
+
+        return response()->json($pivot);
+    }
+
+    public function unlink{{subresourceClass}}($entityId, $relatedEntityId) {
+        if (Gate::denies('delete', new {{subresourceClass}}())) {
+            return self::abort(HttpStatus::Forbidden, [
+                'title' => 'Not allowed to unlink the resource'
+            ]);
+        }
+
+        $entity = {{resourceClass}}::findOrFail($entityId);
+        foreach ($entity->{{subFn}} as $relatedEntity) {
+            if ($relatedEntity->{{subresourceId}} === (int) $relatedEntityId) {
                 $pivot = $relatedEntity->pivot;
                 break;
             }
@@ -88,7 +154,13 @@ class {{resourceClass}}Controller extends RestController {
         return response()->json($pivot);
     }
 
-    public static function unlink{{subresourceClassPlural}}(Request $req, $entityId) {
+    public function unlink{{subresourceClassPlural}}(Request $req, $entityId) {
+        if (Gate::denies('delete', new {{subresourceClass}}())) {
+            return self::abort(HttpStatus::Forbidden, [
+                'title' => 'Not allowed to unlink the resource'
+            ]);
+        }
+
         $entity = {{resourceClass}}::findOrFail($entityId);
         $pivots = [];
 
@@ -105,11 +177,21 @@ class {{resourceClass}}Controller extends RestController {
             }
             $entity->{{subFn}}()->detach($req-all());
         } else {
-            // ...
+            return self::abort(HttpStatus::BadRequest, [
+                'title' => 'Wrong data format',
+                'detail' => 'Either an empty body or a numeric array must be provided'
+            ]);
         }
 
         return response()->json($pivots);
-    }{{else}}public static function delete{{subresourceClass}}($entityId, $relatedEntityId) {
+    }{{else}}
+    public function delete{{subresourceClass}}($entityId, $relatedEntityId) {
+        if (Gate::denies('delete', new {{subresourceClass}}())) {
+            return self::abort(HttpStatus::Forbidden, [
+                'title' => 'Not allowed to delete the resource'
+            ]);
+        }
+
         $entity = {{resourceClass}}::findOrFail($entityId);
         $relatedEntity = $entity->{{subFn}}()->findOrFail($relatedEntityId);
         {{subresourceClass}}::destroy($relatedEntity->{{subresourceId}});
@@ -117,18 +199,28 @@ class {{resourceClass}}Controller extends RestController {
         return response()->json($relatedEntity);
     }
 
-    public static function delete{{subresourceClassPlural}}(Request $req, $entityId) {
-        $entity = {{resourceClass}}::findOrFail($entityId);
+    public function delete{{subresourceClassPlural}}(Request $req, $entityId) {
+        if (Gate::denies('delete', new {{subresourceClass}}())) {
+            return self::abort(HttpStatus::Forbidden, [
+                'title' => 'Not allowed to delete the resource'
+            ]);
+        }
 
-        {{#if isOneToX}}$related = $entity->{{subFn}};
-        $entity->{{subFn}}()->delete();{{else}}if (!$req->getContent()) {
+        $entity = {{resourceClass}}::findOrFail($entityId);
+        {{#if isOneToOne}}
+        $related = $entity->{{subFn}};
+        $entity->{{subFn}}()->delete();{{else}}
+        if (!$req->getContent()) {
             $related = $entity->{{subFn}};
             $entity->{{subFn}}()->delete();
         } else if (is_array($req->all())) {
             $related = $entity->{{subFn}}->findOrFail($req->all());
             {{subresourceClass}}::destroy($req->all());
         } else {
-            // ...
+            return self::abort(HttpStatus::BadRequest, [
+                'title' => 'Wrong data format',
+                'detail' => 'Either an empty body or a numeric array must be provided'
+            ]);
         }{{/if}}
 
         return response()->json($related);
