+++ Engine/src/main/resources/templates/PhpLumen/routes-resource.php
@@ -1,18 +1,24 @@
 // {{resourceName}}
 
-$app->get('/{{resourceRoute}}', '{{resourceClass}}Controller@getMany');
-$app->get('/{{resourceRoute}}/{id}', '{{resourceClass}}Controller@getOne');
+$app->get('/{{resourceRoute}}', ['middleware' => [{{#unless guestCanReadResource}}'auth', {{/unless}}], 'uses' => '{{resourceClass}}Controller@getMany']);
+$app->get('/{{resourceRoute}}/{id}', ['middleware' => [{{#unless guestCanReadResource}}'auth', {{/unless}}], 'uses' => '{{resourceClass}}Controller@getOne']);
 
-$app->post('/{{resourceRoute}}', '{{resourceClass}}Controller@create');
+$app->post('/{{resourceRoute}}', ['middleware' => [{{#unless guestCanCreateResource}}'auth', {{/unless}}], 'uses' => '{{resourceClass}}Controller@create']);
 
-$app->put('/{{resourceRoute}}', '{{resourceClass}}Controller@fullUpdateMany');
-$app->put('/{{resourceRoute}}/{id}', '{{resourceClass}}Controller@fullUpdateOne');
+$app->put('/{{resourceRoute}}', ['middleware' => [{{#unless guestCanUpdateResource}}'auth', {{/unless}}], 'uses' => '{{resourceClass}}Controller@fullUpdateMany']);
+$app->put('/{{resourceRoute}}/{id}', ['middleware' => [{{#unless guestCanUpdateResource}}'auth', {{/unless}}], 'uses' => '{{resourceClass}}Controller@fullUpdateOne']);
 
-$app->patch('/{{resourceRoute}}', '{{resourceClass}}Controller@partialUpdateMany');
-$app->patch('/{{resourceRoute}}/{id}', '{{resourceClass}}Controller@partialUpdateOne');
+$app->patch('/{{resourceRoute}}', ['middleware' => [{{#unless guestCanUpdateResource}}'auth', {{/unless}}], 'uses' => '{{resourceClass}}Controller@partialUpdateMany']);
+$app->patch('/{{resourceRoute}}/{id}', ['middleware' => [{{#unless guestCanUpdateResource}}'auth', {{/unless}}], 'uses' => '{{resourceClass}}Controller@partialUpdateOne']);
 
-$app->delete('/{{resourceRoute}}', '{{resourceClass}}Controller@deleteMany');
-$app->delete('/{{resourceRoute}}/{id}', '{{resourceClass}}Controller@deleteOne');
+$app->delete('/{{resourceRoute}}', ['middleware' => [{{#unless guestCanDeleteResource}}'auth', {{/unless}}], 'uses' => '{{resourceClass}}Controller@deleteMany']);
+$app->delete('/{{resourceRoute}}/{id}', ['middleware' => [{{#unless guestCanDeleteResource}}'auth', {{/unless}}], 'uses' => '{{resourceClass}}Controller@deleteOne']);
+{{#each resourceFiles}}
+$app->get('/{{resourceRoute}}/{id}/files/{{name}}', ['middleware' => [{{#unless guestCanReadResource}}'auth', {{/unless}}], 'uses' => '{{resourceClass}}Controller@get{{camelName}}File']);
+$app->post('/{{resourceRoute}}/{id}/files/{{name}}', ['middleware' => [{{#unless guestCanReadResource}}'auth', {{/unless}}], 'uses' => '{{resourceClass}}Controller@upload{{camelName}}File']);
+$app->put('/{{resourceRoute}}/{id}/files/{{name}}', ['middleware' => [{{#unless guestCanReadResource}}'auth', {{/unless}}], 'uses' => '{{resourceClass}}Controller@upload{{camelName}}File']);
+$app->patch('/{{resourceRoute}}/{id}/files/{{name}}', ['middleware' => [{{#unless guestCanReadResource}}'auth', {{/unless}}], 'uses' => '{{resourceClass}}Controller@upload{{camelName}}File']);
+$app->delete('/{{resourceRoute}}/{id}/files/{{name}}', ['middleware' => [{{#unless guestCanReadResource}}'auth', {{/unless}}], 'uses' => '{{resourceClass}}Controller@delete{{camelName}}File']);{{/each}}
 
 {{#each relations}}{{> routes-relation}}
